Lazy load $zilla->version and other attributes.
The reason for this is, with plugins such as [VersionFromModule], calling `$zilla->version`
in the BeforeBuild phase causes a missing main_module, because file gathering hasn't
happened yet.

Maybe these plugin providers have to be fixed so as it won't die in such cases, but
lazy evaluating these methods in Run does no harm and makes sense to me.